Update Star.pde
Cast to float

+++ Star.pde
@@ -1,9 +1,9 @@
 class Star //note that this class does NOT extend Floater
 {
-  private int myX, myY;
+  private float myX, myY;
   public Star() {
-   myX = (int)(Math.random()*500);
-   myY = (int)(Math.random()*500);
+   myX = (float)(Math.random()*500);
+   myY = (float)(Math.random()*500);
   }
   public void show() {
     color c = color(255, 204, 0); //yellow
